+++ core/server/models/base/bookshelf.js
@@ -62,6 +62,8 @@ ghostBookshelf.plugin(require('./plugins/user-type'));
 
 ghostBookshelf.plugin(require('./plugins/data-manipulation'));
 
+ghostBookshelf.plugin(require('./plugins/overrides'));
+
 // Manages nested updates (relationships)
 ghostBookshelf.plugin('bookshelf-relations', {
     allowedOptions: ['context', 'importing', 'migrating'],
